queue_services/entity-digital-credentials/requirements/prod.txt
sqlalchemy < 1.4.0
Flask
Flask-Babel
Flask-Migrate
Flask-Moment
Flask-Script
Flask-SQLAlchemy
asyncio-nats-client
asyncio-nats-streaming
attrs==23.1.0
jinja2
python-dotenv<0.16.0
psycopg2-binary
requests
sentry-sdk[flask]
Werkzeug<1.0.2
